Added custom object loader and custom texture loaders
I went and found an open source object loader (I take no credit for that) and added it, and I've kept the original file names as welll.
I've added custom texture loading so people can add their own images into the game. Fixed an issue with ore generation breaking when no custom ore was added to the game (A for loop didn't have a proper check when the list was empty).